pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkup-cli"
version = "0.3.1"
description = "Official CLI for Linkup - AI-powered web search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Linkup", email = "support@linkup.so" }
]
keywords = ["linkup", "cli", "search", "ai", "web-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
    "linkup-sdk>=0.2.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://linkup.so"
Documentation = "https://docs.linkup.so"
Repository = "https://github.com/shauryajain21/linkup-cli"
Issues = "https://github.com/shauryajain21/linkup-cli/issues"

[project.scripts]
linkup = "linkup_cli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/linkup_cli"]
